+++ demo/src/plugins.rs
@@ -1,6 +1,6 @@
 use egui::{Color32, Response, Ui};
 use walkers::{
-    extras::{Image, Images, Place, Places, Style, Texture},
+    extras::{Image, LabeledSymbol, LabeledSymbolStyle, Places, Texture},
     Plugin, Position, Projector,
 };
 
@@ -9,17 +9,17 @@ use crate::places;
 /// Creates a built-in `Places` plugin with some predefined places.
 pub fn places() -> impl Plugin {
     Places::new(vec![
-        Place {
+        LabeledSymbol {
             position: places::wroclaw_glowny(),
             label: "Wrocław Główny\ntrain station".to_owned(),
             symbol: '🚆',
-            style: Style::default(),
+            style: LabeledSymbolStyle::default(),
         },
-        Place {
+        LabeledSymbol {
             position: places::dworcowa_bus_stop(),
             label: "Bus stop".to_owned(),
             symbol: '🚌',
-            style: Style::default(),
+            style: LabeledSymbolStyle::default(),
         },
     ])
 }
@@ -45,7 +45,7 @@ impl ImagesPluginData {
 
 /// Creates a built-in `Images` plugin with an example image.
 pub fn images(images_plugin_data: &mut ImagesPluginData) -> impl Plugin {
-    Images::new(vec![{
+    Places::new(vec![{
         let mut image = Image::new(images_plugin_data.texture.clone(), places::wroclavia());
         image.scale(images_plugin_data.x_scale, images_plugin_data.y_scale);
         image.angle(images_plugin_data.angle.to_radians());