+++ walkers/src/extras/mod.rs
@@ -1,6 +1,4 @@
 //! Extra functionalities that can be used with the map.
 mod places;
-pub use places::{Place, Places, Style};
-mod images;
 pub use crate::tiles::Texture;
-pub use images::{Image, Images};
+pub use places::{Image, LabeledSymbol, LabeledSymbolStyle, Places};