+++ walkers/src/extras/places.rs
@@ -1,35 +1,43 @@
-use egui::{vec2, Align2, Color32, FontId, Response, Stroke, Ui};
+use egui::{emath::Rot2, vec2, Align2, Color32, FontId, Rect, Response, Stroke, Ui, Vec2};
 
 use crate::{Plugin, Position};
 
-/// Visual style of the place.
-#[derive(Clone)]
-pub struct Style {
-    pub label_font: FontId,
-    pub label_color: Color32,
-    pub label_background: Color32,
-    pub symbol_font: FontId,
-    pub symbol_color: Color32,
-    pub symbol_background: Color32,
-    pub symbol_stroke: Stroke,
+use super::Texture;
+
+/// [`Plugin`] which draws list of places on the map.
+pub struct Places<T>
+where
+    T: Place,
+{
+    places: Vec<T>,
 }
 
-impl Default for Style {
-    fn default() -> Self {
-        Self {
-            label_font: FontId::proportional(12.),
-            label_color: Color32::from_gray(200),
-            label_background: Color32::BLACK.gamma_multiply(0.8),
-            symbol_font: FontId::proportional(14.),
-            symbol_color: Color32::BLACK.gamma_multiply(0.8),
-            symbol_background: Color32::WHITE.gamma_multiply(0.8),
-            symbol_stroke: Stroke::new(2., Color32::BLACK.gamma_multiply(0.8)),
+impl<T> Places<T>
+where
+    T: Place,
+{
+    pub fn new(places: Vec<T>) -> Self {
+        Self { places }
+    }
+}
+
+impl<T> Plugin for Places<T>
+where
+    T: Place + 'static,
+{
+    fn run(self: Box<Self>, ui: &mut Ui, _response: &Response, projector: &crate::Projector) {
+        for place in &self.places {
+            place.draw(ui, projector);
         }
     }
 }
 
-/// A place to be drawn on the map.
-pub struct Place {
+pub trait Place {
+    fn draw(&self, ui: &Ui, projector: &crate::Projector);
+}
+
+/// A symbol with a label to be drawn on the map.
+pub struct LabeledSymbol {
     /// Geographical position.
     pub position: Position,
 
@@ -41,10 +49,10 @@ pub struct Place {
     pub symbol: char,
 
     /// Visual style of this place.
-    pub style: Style,
+    pub style: LabeledSymbolStyle,
 }
 
-impl Place {
+impl Place for LabeledSymbol {
     fn draw(&self, ui: &Ui, projector: &crate::Projector) {
         let screen_position = projector.project(self.position);
         let painter = ui.painter();
@@ -87,21 +95,76 @@ impl Place {
     }
 }
 
-/// [`Plugin`] which draws list of places on the map.
-pub struct Places {
-    places: Vec<Place>,
+/// Visual style of the place.
+#[derive(Clone)]
+pub struct LabeledSymbolStyle {
+    pub label_font: FontId,
+    pub label_color: Color32,
+    pub label_background: Color32,
+    pub symbol_font: FontId,
+    pub symbol_color: Color32,
+    pub symbol_background: Color32,
+    pub symbol_stroke: Stroke,
 }
 
-impl Places {
-    pub fn new(places: Vec<Place>) -> Self {
-        Self { places }
+impl Default for LabeledSymbolStyle {
+    fn default() -> Self {
+        Self {
+            label_font: FontId::proportional(12.),
+            label_color: Color32::from_gray(200),
+            label_background: Color32::BLACK.gamma_multiply(0.8),
+            symbol_font: FontId::proportional(14.),
+            symbol_color: Color32::BLACK.gamma_multiply(0.8),
+            symbol_background: Color32::WHITE.gamma_multiply(0.8),
+            symbol_stroke: Stroke::new(2., Color32::BLACK.gamma_multiply(0.8)),
+        }
     }
 }
 
-impl Plugin for Places {
-    fn run(self: Box<Self>, ui: &mut Ui, _response: &Response, projector: &crate::Projector) {
-        for place in &self.places {
-            place.draw(ui, projector);
+/// An image to be drawn on the map.
+pub struct Image {
+    /// Geographical position.
+    position: Position,
+
+    scale: Vec2,
+    angle: Rot2,
+    texture: Texture,
+}
+
+impl Image {
+    pub fn new(texture: Texture, position: Position) -> Self {
+        Self {
+            position,
+            scale: Vec2::splat(1.0),
+            angle: Rot2::from_angle(0.0),
+            texture,
+        }
+    }
+
+    /// Scale the image.
+    pub fn scale(&mut self, x: f32, y: f32) {
+        self.scale.x = x;
+        self.scale.y = y;
+    }
+
+    /// Set the image's angle in radians.
+    pub fn angle(&mut self, angle: f32) {
+        self.angle = Rot2::from_angle(angle);
+    }
+}
+
+impl Place for Image {
+    fn draw(&self, ui: &Ui, projector: &crate::Projector) {
+        let painter = ui.painter();
+        let rect = Rect::from_center_size(
+            projector.project(self.position).to_pos2(),
+            self.texture.size() * self.scale,
+        );
+
+        if painter.clip_rect().intersects(rect) {
+            let mut mesh = self.texture.mesh_with_rect(rect);
+            mesh.rotate(self.angle, rect.center());
+            painter.add(mesh);
         }
     }
 }